# Bressolium/backend/BressoliumProject/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bressolium API",
        "description": "API REST del videojuego de estrategia colaborativa Bressolium. Todos los endpoints devuelven JSON con el formato unificado {success, data, error}.",
        "contact": {
            "name": "Equipo Bressolium"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/game/{gameId}/vote": {
            "post": {
                "tags": [
                    "Vote"
                ],
                "summary": "Vota una tecnología o un invento en la jornada actual",
                "description": "El usuario vota por la tecnología o invento que quiere investigar/construir. Si no se envía ni technology_id ni invention_id se considera abstención. Solo se puede votar una vez por jornada.",
                "operationId": "1639536305c5cdbed4489448db130ed5",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "UUID de la partida",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "technology_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "01HFG...",
                                        "nullable": true
                                    },
                                    "invention_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voto registrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "vote_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No perteneces a esta partida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partida no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "description": "Respuesta de error estándar",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "example": "Mensaje de error"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de Laravel Sanctum obtenido al hacer login. Se envía en la cabecera Authorization: Bearer {token}.",
                "bearerFormat": "Sanctum-Token",
                "scheme": "bearer"
            }
        }
    },
...
}